fix(common): http service should not cancel when response type is stream
when response type stream is set, the node stream is ended early as `subscriber.complete()` -> `cancelSource.cancel()` is called in sequence.
The better approach to fix this is to wrap the stream with an observable, but that'll be a lot of code.
So a simpler approach is to ignore request cancellation on stream response type.

fixes #4014